[refactor] 제미나이 리뷰 코드 작성 마무리
